유니모 성장 시 특화 스탯 가중치 추가 (LocalPlayer.cs에 레벨 변수 수정 필요)

+++ Assets/00_TeamMaple/Scripts/SO/Unimo/UnimoStatLevelUpDataSO.cs
@@ -10,7 +10,7 @@ public class UnimoStatLevelUpData
     public int Level { get; set; }
     public int PlusHp { get; set; }
     public int PlusDef { get; set; }
-    public int PlusSpeed { get; set; }
+    public float PlusSpeed { get; set; }
     public int PlusBloomRange { get; set; }
     public float PlusBloomSpeed { get; set; }
     public float PlusFlowerRate { get; set; }
